refactor(web/technology-form): create a helper for reuse funding data from review
